# Faster (and safer) equal implementation