@clauderic 2022-04-01
Hey @tmcw, thanks for your contribution! This looks great! Would you mind [adding a changeset](https://github.com/placemark/dnd-kit/new/faster-equal?filename=.changeset/shaggy-masks-refuse.md&value=---%0A%22%40dnd-kit%2Fsortable%22%3A%20patch%0A---%0A%0AFaster%20(and%20safer)%20equal%20implementation%0A) or giving me write access?