@tmcw 2022-05-03
👋  hey - any chance that this might get released? Still seeing this in my profiles and very excited to remove that bottleneck