@clauderic 2022-05-03
Hey @tmcw, every merged commit gets deployed to npm under the `next` tag, so you can already upgrade to the packages of @dnd-kit tagged with `next`:

```bash
npm install @dnd-kit/core@next
```